branch: 3 description: - ajustes no código para lista com nomes em branco.
